organize SignupView
