+++ src/views/SignupView/StudentForm.vue
@@ -193,7 +193,7 @@
 import validator from 'validator'
 
 import AuthService from 'src/services/AuthService'
-import UserService from '../../services/UserService'
+import UserService from 'src/services/UserService'
 
 export default {
   data () {
